configure.in: don't force use of cache
Back when hardware was slower and we had lots and lots of
imakefile-like configure scripts, forcing the use of config.cache made
sense for performance reasons. Now that we have one configure script
and we build on modern hardware, config.cache just introduces errors
as you are debugging changes to the configure script.

This commit stops second guessing autoconf's cache handling.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@24468 dc483132-0cff-0310-8789-dd5450dbe970