ticket: 6941 subject: Fix accidental KDC use of replay cache target_version: 1.9.2 tags: pullup
r24464 (ticket #6804) intended to remove the KDC replay cache by
eliminating all of the USE_RCACHE code, but it had the unintended side
effect of causing krb5_rd_req_decoded to use the default server
rcache.  Using this cache is much less efficient because it is opened
and re-read for each request.

Set appropriate flags on the auth context to disable replay cache use
for TGS requests altogether.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25076 dc483132-0cff-0310-8789-dd5450dbe970