Use hard links for LIBLINKS.
verto uses dladdr() to figure out where it should look for modules,
and dladdr may or may not follow symlinks.  Hard links ensure that it
always looks in the directory where liblinks are, allowing it to find
the verto-k5ev module.

Make symlinks within the lib directory for .so and .so.X instead of
symlinking those back to the library build directory.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25150 dc483132-0cff-0310-8789-dd5450dbe970