ticket: 6964 subject: Support special salt type in default krb5_dbe_cpw.
This change allows the "special" salt type to be used in
supported_enctypes or in the argument to kadmin's cpw -e.  If used,
kadmind will pick a salt consisting of 64 random bits represented as
16 printable ASCII characters.  The use of random explicit salts
creates some interoperability issues and is not generally recommended,
but can be useful for interop testing, as a workaround for obscure
bugs, or to increase the difficulty of brute-force password searches
in situations where none of the interoperability issues apply.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25226 dc483132-0cff-0310-8789-dd5450dbe970